cleanup callers & move timestamp parsing to utils
